got rid of the firebaseservice file
